+++ src/app/(main)/(mobile)/me/layout.tsx
@@ -2,8 +2,14 @@ import { PropsWithChildren } from 'react';
 
 import MobileContentLayout from '@/components/server/MobileNavLayout';
 
+import Header from './features/Header';
+
 const Layout = ({ children }: PropsWithChildren) => {
-  return <MobileContentLayout withNav>{children}</MobileContentLayout>;
+  return (
+    <MobileContentLayout header={<Header />} withNav>
+      {children}
+    </MobileContentLayout>
+  );
 };
 
 Layout.displayName = 'MeLayout';